Safari-611.1.10/LayoutTests/contact-picker/contacts-select-requires-user-gesture-expected.txt

This test verifies that navigator.contacts.select requires a user gesture.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Did not present contact picker.
PASS exception.name is "SecurityError"
PASS successfullyParsed is true

TEST COMPLETE
